+++ UploadDaemon/SymbolAnalysis/ParsedTraceFile.cs
@@ -2,7 +2,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
+using System.Resources;
 using System.Text.RegularExpressions;
+using static UploadDaemon.SymbolAnalysis.RevisionFileUtils;
 
 namespace UploadDaemon.SymbolAnalysis
 {
@@ -33,6 +36,16 @@ namespace UploadDaemon.SymbolAnalysis
         private static readonly Regex AssemblyLineRegex = new Regex(@"^Assembly=(?<name>[^:]+):(?<id>\d+).*?(?: Path:(?<path>.*))?$");
         private static readonly Regex CoverageLineRegex = new Regex(@"^(?:Inlined|Jitted)=(\d+):(?:\d+:)?(\d+)");
 
+        /// <summary>
+        /// The name of the Resource .resx file that holed information about embedded upload targets.
+        /// </summary>
+        private const String TeamscaleResourceName = "Teamscale";
+
+        /// <summary>
+        /// The uploads targets (revision/timestamp and optionally teamscale project) that are retrieved from resource files that are embedded into assemblies referenced in the trace file.
+        /// </summary>
+        public readonly List<(string project, RevisionOrTimestamp revisionOrTimestamp)> embeddedUploadTargets = new List<(string project, RevisionOrTimestamp revisionOrTimestamp)>();
+
         public ParsedTraceFile(string[] lines, string filePath)
         {
             this.FilePath = filePath;
@@ -44,7 +57,7 @@ namespace UploadDaemon.SymbolAnalysis
                     match => (name: match.Groups["name"].Value, path: match.Groups["path"].Value)
                 );
             this.LoadedAssemblies = assemblyTokens.Values.ToList();
-
+            SearchForEmbeddedUploadTargets();
             IEnumerable<Match> coverageMatches = lines.Select(line => CoverageLineRegex.Match(line))
                             .Where(match => match.Success);
             foreach (Match match in coverageMatches)
@@ -58,6 +71,89 @@ namespace UploadDaemon.SymbolAnalysis
                 }
                 CoveredMethods.Add((assembly.name, Convert.ToUInt32(match.Groups[2].Value)));
             }
+        }
+
+        /// <summary>
+        /// Checks the loaded assemblies for resources that contain information about target revision or teamscale projects.
+        /// </summary>
+        private void SearchForEmbeddedUploadTargets()
+        {
+            foreach ((_, string path) in this.LoadedAssemblies)
+            {
+                Assembly assembly = LoadAssemblyFromPath(path);
+                if (assembly == null || assembly.DefinedTypes == null)
+                {
+                    continue;
+                }
+                TypeInfo teamscaleResourceType = assembly.DefinedTypes.FirstOrDefault(x => x.Name == TeamscaleResourceName) ?? null;
+                if (teamscaleResourceType == null)
+                {
+                    continue;
+                }
+                logger.Info("Found embedded Teamscale resource in {assembly} that can be used to identify upload targets.", assembly);
+                ResourceManager teamscaleResourceManager = new ResourceManager(teamscaleResourceType.FullName, assembly);
+                string embeddedTeamscaleProject = teamscaleResourceManager.GetString("Project");
+                string embeddedRevision = teamscaleResourceManager.GetString("Revision");
+                string embeddedTimestamp = teamscaleResourceManager.GetString("Timestamp");
+                AddUploadTarget(embeddedRevision, embeddedTimestamp, embeddedTeamscaleProject, embeddedUploadTargets, assembly.FullName);
+            }
+        }
+
+        /// <summary>
+        /// Adds a revision or timestamp and optionally a project to the list of upload targets. This method checks if both, revision and timestamp, are declared, or neither.
+        /// </summary>
+        /// <param name="revision"></param>
+        /// <param name="timestamp"></param>
+        /// <param name="project"></param>
+        /// <param name="uploadTargets"></param>
+        /// <param name="origin"></param>
+        public static void AddUploadTarget(string revision, string timestamp, string project, List<(string project, RevisionOrTimestamp revisionOrTimestamp)> uploadTargets, string origin)
+        {
+            Logger logger = LogManager.GetCurrentClassLogger();
+
+            if (revision == null && timestamp == null)
+            {
+                logger.Error("Not all required fields in {origin}. Please specify either 'Revision' or 'Timestamp'", origin);
+                return;
+            }
+            if (revision != null && timestamp != null)
+            {
+                logger.Error("'Revision' and 'Timestamp' are both set in {origin}. Please set only one, not both.", origin);
+                return;
+            }
+            if (revision != null)
+            {
+                uploadTargets.Add((project, new RevisionOrTimestamp(revision, true)));
+            }
+            else
+            {
+                uploadTargets.Add((project, new RevisionOrTimestamp(timestamp, false)));
+            }
+        }
+
+        private Assembly LoadAssemblyFromPath(string path)
+        {
+            if (String.IsNullOrEmpty(path))
+            {
+                return null;
+            }
+            Assembly assembly;
+            try
+            {
+                assembly = Assembly.LoadFrom(path);
+                // Check that defined types can actually be loaded
+                if (assembly == null)
+                {
+                    return null;
+                }
+                IEnumerable<TypeInfo> ignored = assembly.DefinedTypes;
+            }
+            catch (Exception e)
+            {
+                logger.Debug("Could not load {assembly}. Skipping upload resource discovery. {e}", path, e);
+                return null;
+            }
+            return assembly;
         }
     }
 }