+++ UploadDaemon/SymbolAnalysis/RevisionFileUtils.cs
@@ -16,6 +16,12 @@ namespace UploadDaemon.SymbolAnalysis
         /// </summary>
         public class RevisionOrTimestamp
         {
+            public RevisionOrTimestamp(string value, bool isRevision)
+            {
+                this.Value = value;
+                this.IsRevision = isRevision;
+            }
+
             /// <summary>
             /// The timestamp or revision.
             /// </summary>
@@ -66,23 +72,15 @@ namespace UploadDaemon.SymbolAnalysis
                     " found neither a timestamp nor a revision entry." +
                     " Examples: 'timestamp: 1234567890' or 'revision: 123456'");
             }
-
             (string type, string value) = matches.First();
+
             switch (type.ToLower())
             {
                 case "timestamp":
-                    return new RevisionOrTimestamp
-                    {
-                        Value = value,
-                        IsRevision = false,
-                    };
+                    return new RevisionOrTimestamp(value, false);
 
                 case "revision":
-                    return new RevisionOrTimestamp
-                    {
-                        Value = value,
-                        IsRevision = true,
-                    };
+                    return new RevisionOrTimestamp(value, true);
 
                 default:
                     throw new InvalidRevisionFileException($"The revision file {filePath} is not valid:" +